Adding test classes and cleaning and reformatting all code.
